requirements.txt
Django==4.2
djangorestframework==3.15.0
djangorestframework-simplejwt==5.2.2
PyMySQL==1.1.0
python-dotenv==1.0.0
